Fix clang error

+++ rts/Rendering/Textures/S3OTextureHandler.cpp
@@ -97,7 +97,7 @@ int CS3OTextureHandler::LoadS3OTextureNow(const S3DModel* model)
 		if (model->invertTexYAxis)
 			texBitMaps[0].ReverseYAxis();
 
-		textureCache[model->tex1] = {texBitMaps[0].CreateTexture(true), texBitMaps[0].xsize, texBitMaps[0].ysize};
+		textureCache[model->tex1] = {texBitMaps[0].CreateTexture(true), (unsigned int)texBitMaps[0].xsize, (unsigned int)texBitMaps[0].ysize};
 	}
 
 	if (texCacheIters[1] == textureCache.end()) {
@@ -114,7 +114,7 @@ int CS3OTextureHandler::LoadS3OTextureNow(const S3DModel* model)
 		if (model->invertTexYAxis)
 			texBitMaps[1].ReverseYAxis();
 
-		textureCache[model->tex2] = {texBitMaps[1].CreateTexture(true), texBitMaps[1].xsize, texBitMaps[1].ysize};
+		textureCache[model->tex2] = {texBitMaps[1].CreateTexture(true), (unsigned int)texBitMaps[1].xsize, (unsigned int)texBitMaps[1].ysize};
 	}
 
 	if (texCacheIters[0] == textureCache.end() || texCacheIters[1] == textureCache.end()) {
